Update to qmk cli and qmk_firmware 0.14.29

+++ scripts/build.sh
@@ -1,14 +1,16 @@
-QMK_SRC_DIR=/qmk
-QMK_KEYMAPS_DIR=${QMK_SRC_DIR}/keyboards/ergodox_ez/keymaps
-MYKEYMAP_DIR=/keymap
-MYKEYMAP_NAME=TribalNightOwl
-KEYBOARD=ergodox_ez
+#!/usr/bin/env bash
 
-rm -rf ${QMK_KEYMAPS_DIR}/${MYKEYMAP_NAME}
-cp -r  ${MYKEYMAP_DIR} ${QMK_KEYMAPS_DIR}/${MYKEYMAP_NAME}
+QMK_SRC_DIR="/qmk"
+QMK_KEYMAPS_DIR="${QMK_SRC_DIR}/keyboards/ergodox_ez/keymaps"
+MYKEYMAP_DIR="/keymap"
+MYKEYMAP_NAME="TribalNightOwl"
+KEYBOARD="ergodox_ez"
 
-cd ${QMK_SRC_DIR}
+rm -rf "${QMK_KEYMAPS_DIR:?}/${MYKEYMAP_NAME:?}"
+cp -r  "${MYKEYMAP_DIR}" "${QMK_KEYMAPS_DIR}/${MYKEYMAP_NAME}"
 
+cd "${QMK_SRC_DIR}" || exit
 
 make git-submodule
-make ${KEYBOARD}:${MYKEYMAP_NAME}
+
+qmk compile -kb "${KEYBOARD}" -km "${MYKEYMAP_NAME}"
